backend/app/requirements.txt
gradio
ollama
weaviate-client==4.7.0rc2
requests
